cmd/server: move search logic out of routes for re-use
Periodic searches need to use the same logic as the HTTP routes do, so
we need to centalize that logic in functions.
