+++ cmd/server/search_handlers.go
@@ -1,0 +1,154 @@
+// Copyright 2018 The Moov Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
+
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"net/url"
+	"strings"
+
+	moovhttp "github.com/moov-io/base/http"
+
+	"github.com/go-kit/kit/log"
+	"github.com/gorilla/mux"
+)
+
+func addSearchRoutes(logger log.Logger, r *mux.Router, searcher *searcher) {
+	r.Methods("GET").Path("/search").HandlerFunc(search(logger, searcher))
+}
+
+type addressSearchRequest struct {
+	Address    string `json:"address"`
+	City       string `json:"city"`
+	State      string `json:"state"`
+	Providence string `json:"providence"`
+	Zip        string `json:"zip"`
+	Country    string `json:"country"`
+}
+
+func (req addressSearchRequest) empty() bool {
+	return req.Address == "" && req.City == "" && req.State == "" &&
+		req.Providence == "" && req.Zip == "" && req.Country == ""
+}
+
+func readAddressSearchRequest(u *url.URL) addressSearchRequest {
+	return addressSearchRequest{
+		Address:    strings.ToLower(strings.TrimSpace(u.Query().Get("address"))),
+		City:       strings.ToLower(strings.TrimSpace(u.Query().Get("city"))),
+		State:      strings.ToLower(strings.TrimSpace(u.Query().Get("state"))),
+		Providence: strings.ToLower(strings.TrimSpace(u.Query().Get("providence"))),
+		Zip:        strings.ToLower(strings.TrimSpace(u.Query().Get("zip"))),
+		Country:    strings.ToLower(strings.TrimSpace(u.Query().Get("country"))),
+	}
+}
+
+func search(logger log.Logger, searcher *searcher) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		w, err := wrapResponseWriter(logger, w, r)
+		if err != nil {
+			return
+		}
+
+		// Search by Name
+		if name := strings.TrimSpace(r.URL.Query().Get("name")); name != "" {
+			if logger != nil {
+				logger.Log("search", fmt.Sprintf("searching SDN names for %s", name))
+			}
+			searchByName(logger, searcher, name)(w, r)
+			return
+		}
+
+		// Search by Alt Name
+		if alt := strings.TrimSpace(r.URL.Query().Get("altName")); alt != "" {
+			if logger != nil {
+				logger.Log("search", fmt.Sprintf("searching SDN alt names for %s", alt))
+			}
+			searchByAltName(logger, searcher, alt)(w, r)
+			return
+		}
+
+		// Search Addresses
+		if req := readAddressSearchRequest(r.URL); !req.empty() {
+			if logger != nil {
+				logger.Log("search", fmt.Sprintf("searching address for %#v", req))
+			}
+			searchByAddress(logger, searcher, req)(w, r)
+			return
+		}
+
+		// Fallback if no search params were found
+		moovhttp.Problem(w, errNoSearchParams)
+	}
+}
+
+func searchByAddress(logger log.Logger, searcher *searcher, req addressSearchRequest) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		hasAddress := req.Address != ""
+		reqAdds := strings.Fields(strings.ToLower(req.Address))
+
+		if !hasAddress {
+			w.WriteHeader(http.StatusBadRequest)
+			return
+		}
+
+		limit := extractSearchLimit(r)
+		answers := searcher.FindAddresses(limit, func(add *Address) bool {
+			return addressMatches(reqAdds, add)
+		})
+
+		w.Header().Set("Content-Type", "application/json; charset=utf-8")
+		w.WriteHeader(http.StatusOK)
+		if err := json.NewEncoder(w).Encode(&searchResponse{Addresses: answers}); err != nil {
+			moovhttp.Problem(w, err)
+			return
+		}
+	}
+}
+
+func searchByName(logger log.Logger, searcher *searcher, nameSlug string) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		nameSlugs := strings.Fields(strings.ToLower(nameSlug))
+		if len(nameSlugs) == 0 {
+			moovhttp.Problem(w, errNoSearchParams)
+			return
+		}
+
+		limit := extractSearchLimit(r)
+		sdns := searcher.FindSDNs(limit, func(sdn *SDN) bool {
+			return nameMatches(nameSlugs, sdn)
+		})
+
+		w.Header().Set("Content-Type", "application/json; charset=utf-8")
+		w.WriteHeader(http.StatusOK)
+		if err := json.NewEncoder(w).Encode(&searchResponse{SDNs: sdns}); err != nil {
+			moovhttp.Problem(w, err)
+			return
+		}
+	}
+}
+
+func searchByAltName(logger log.Logger, searcher *searcher, altSlug string) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		altSlugs := strings.Fields(strings.ToLower(altSlug))
+		if len(altSlugs) == 0 {
+			moovhttp.Problem(w, errNoSearchParams)
+			return
+		}
+
+		limit := extractSearchLimit(r)
+		alts := searcher.FindAlts(limit, func(alt *Alt) bool {
+			return altMatches(altSlugs, alt)
+		})
+
+		w.Header().Set("Content-Type", "application/json; charset=utf-8")
+		w.WriteHeader(http.StatusOK)
+		if err := json.NewEncoder(w).Encode(&searchResponse{AltNames: alts}); err != nil {
+			moovhttp.Problem(w, err)
+			return
+		}
+	}
+}
